Add NONE permutation option to produce dummy for testing; add cuSPARSE template helpers; fix FAST_MMIO mtx writing bug

+++ examples/benchmarks/smax_benchmarks/smax_crs_cuda_spmv_bench.cpp
@@ -4,8 +4,8 @@
 #include "smax_benchmarks_common.hpp"
 
 // Set datatypes
-using IT = unsigned int;
-using VT = double;
+using IT = int;
+using VT = float;
 
 int main(int argc, char *argv[]) {
